backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Discord bot
discord.py==2.3.2
Pillow==10.2.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9

# AI and ML
pydantic-ai==0.0.13
openai==1.10.0
anthropic==0.18.0

# Real-time
python-socketio==5.11.0
aiohttp==3.9.1

# Utilities
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14

